pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-pack"
version = "0.0.1"
description = "AI-tool-agnostic code review frameworks"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
code-review-pack = "code_review_pack.cli:main"

# [project.urls]
# Homepage = "https://github.com/YOUR-ORG/code-review-packs"
# Documentation = "https://github.com/YOUR-ORG/code-review-packs/docs"
# Repository = "https://github.com/YOUR-ORG/code-review-packs"
# TODO: Uncomment and update URLs before publishing to PyPI

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
code_review_pack = ["templates/**/*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
